html/semantics/forms/the-input-element/color.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Form input type=color</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.whatwg.org/specs/web-apps/current-work/multipage/common-microsyntaxes.html#colors">
<link rel=help href="http://www.whatwg.org/specs/web-apps/current-work/multipage/states-of-the-type-attribute.html#color-state-(type=color)">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  var colors = [
    {value: "", expected: "#000000", testname: "Empty value should return #000000"},
    {value: "#ffffff", expected: "#ffffff", testname: "Valid simple color: should return #ffffff"},
    {value: "#FFFFFF", expected: "#ffffff", testname: "Valid simple color (containing LATIN CAPITAL LETTERS): should return #ffffff (converted to ASCII lowercase)"},
    {value: "#fff", expected: "#000000", testname: "Invalid simple color: not 7 characters long"},
    {value: "fffffff", expected: "#000000", testname: "Invalid simple color: no starting # sign"},
    {value: "#gggggg", expected: "#000000", testname: "Invalid simple color: non ASCII hex digits"},
    {value: "foobar", expected: "#000000", testname: "Invalid simple color: foobar"}
  ];
  for (var i = 0; i < colors.length; i++) {
    var w = colors[i];
    test(function() {
      var input = document.createElement("input");
      input.type = "color";
      input.value = w.value;
      assert_equals(input.value, w.expected);
    }, w.testname);
  }
</script>
